fix(stella-model): repair dropped brace + split zai tests under the ratchet
A main-merge into this branch dropped the closing brace of
zai_identity_never_sends_reasoning_effort_even_when_pinned, leaving
zai/tests.rs with an unclosed delimiter (red 'fmt + clippy + test').
The #266/#268 additions also pushed the file to 1568 lines, over the
1500 cap. Restore the brace and extract the issue #271/#250 error-
classification block into a zai/tests/error_classify_tests.rs submodule;
tests.rs drops to 1440 lines. 217 stella-model tests pass, clippy clean.
